+++ js/home.js
@@ -1,90 +1,5 @@
 import postApi from './api/postApi';
-import { getUlPagination, setTextContent, truncateText } from './utils';
-import dayjs from 'dayjs';
-import relativeTime from 'dayjs/plugin/relativeTime';
-import debounce from 'lodash.debounce';
-
-// To use fromNow function
-dayjs.extend(relativeTime);
-
-function createPostElement(post) {
-  if (!post) return;
-
-  // Find and clone template
-  const postTemplate = document.getElementById('postTemplate');
-  if (!postTemplate) return;
-
-  const liElement = postTemplate.content.firstElementChild.cloneNode(true);
-  if (!liElement) return;
-
-  // Update title, description, author and thumbnail
-  // const titleElement = liElement.querySelector('[data-id="title"]');
-  // if (titleElement) titleElement.textContent = post.title
-
-  // const descriptionElement = liElement.querySelector('[data-id="description"]');
-  // if (descriptionElement) descriptionElement.textContent = post.description
-
-  // const authorElement = liElement.querySelector('[data-id="author"]');
-  // if (authorElement) authorElement.textContent = post.author
-
-  setTextContent(liElement, '[data-id="title"]', post.title);
-  setTextContent(liElement, '[data-id="description"]', truncateText(post.description, 120));
-  setTextContent(liElement, '[data-id="author"]', post.author);
-
-  // Calculate timespan
-  // setTextContent(liElement, '[data-id="timeSpan"]', ` - ${dayjs(post.updatedAt).fromNow()}`);
-  setTextContent(
-    liElement, '[data-id="timeSpan"]', ` - ${dayjs(post.updatedAt).format('DD/MM/YY - h:mm A')}`
-  );
-
-  const thumbnailElement = liElement.querySelector('[data-id="thumbnail"]');
-  if (thumbnailElement) {
-    thumbnailElement.src = post.imageUrl;
-
-    thumbnailElement.addEventListener('error', () => {
-      console.log('Load image failed --> use default placeholder');
-      thumbnailElement.src = 'https://via.placeholder.com/1368x400?text=thumbnail';
-    });
-  }
-
-  // Attach events
-  return liElement;
-}
-
-function renderPostList(postList) {
-  if (!Array.isArray(postList)) return;
-
-  const ulElement = document.getElementById('postList');
-  if (!ulElement) return;
-
-  // Clear current list before render new list
-  ulElement.textContent = '';
-
-  postList.forEach((post) => {
-    const liElement = createPostElement(post);
-    ulElement.appendChild(liElement);
-  });
-}
-
-function renderPagination(pagination) {
-  const ulPagination = getUlPagination();
-  if (!pagination || !ulPagination) return;
-
-  // Calc totalPages
-  const { _page, _limit, _totalRows } = pagination;
-  const totalPages = Math.ceil(_totalRows / _limit);
-
-  // Save page and totalPages to ulPagination
-  ulPagination.dataset.page = _page;
-  ulPagination.dataset.totalPages = totalPages;
-
-  // Check if enable/disable prev/next links
-  if (_page <= 1) ulPagination.firstElementChild?.classList.add('disabled');
-  else ulPagination.firstElementChild?.classList.remove('disabled');
-
-  if (_page >= totalPages) ulPagination.lastElementChild?.classList.add('disabled');
-  else ulPagination.lastElementChild?.classList.remove('disabled');
-}
+import { initPagination, initSearch, renderPostList, renderPagination } from './utils';
 
 async function handleFilterChange(filterName, filterValue) {
   try {
@@ -99,96 +14,45 @@ async function handleFilterChange(filterName, filterValue) {
 
     // Fetch API and re-render post list
     const { data, pagination } = await postApi.getAll(url.searchParams);
-    renderPostList(data);
-    renderPagination(pagination);
+
+    renderPostList('postList', data);
+    renderPagination('pagination', pagination);
   } catch (error) {
     console.log('Failed to fetch post list', error);
   }
 }
 
-function handlePrevClick(event) {
-  event.preventDefault();
-
-  const ulPagination = getUlPagination();
-  if (!ulPagination) return;
-
-  // Convert (typeof page) string to number
-  const page = Number.parseInt(ulPagination.dataset.page) || 1;
-  if (page <= 1) return;
-
-  handleFilterChange('_page', page - 1);
-}
-
-function handleNextClick(event) {
-  event.preventDefault();
-
-  const ulPagination = getUlPagination();
-  if (!ulPagination) return;
-
-  const page = Number.parseInt(ulPagination.dataset.page) || 1;
-  const totalPages = ulPagination.dataset.totalPages;
-  if (page >= totalPages) return;
-
-  handleFilterChange('_page', page + 1);
-}
-
-function initPagination() {
-  // Bind click event for prev/next link (ul > li > a)
-  const ulPagination = getUlPagination();
-  if (!ulPagination) return;
-
-  // Add click event for prev link
-  const prevLink = ulPagination.firstElementChild?.firstElementChild;
-  if (prevLink) {
-    prevLink.addEventListener('click', handlePrevClick);
-  }
-
-  // Add click event for next link
-  const nextLink = ulPagination.lastElementChild?.lastElementChild;
-  if (nextLink) {
-    nextLink.addEventListener('click', handleNextClick);
-  }
-}
-
-function initURL() {
-  const url = new URL(window.location);
-
-  // Update search params if needed
-  if (!url.searchParams.get('_page')) url.searchParams.set('_page', 1);
-  if (!url.searchParams.get('_limit')) url.searchParams.set('_limit', 6);
-  history.pushState({}, '', url);
-}
-
-function initSearch() {
-  const searchInput = document.getElementById('searchInput');
-  if (!searchInput) return;
-
-  // Set default values from query params
-  const queryParams = new URLSearchParams(window.location.search);
-  if (queryParams.get('title_like')) {
-    searchInput.value = queryParams.get('title_like');
-  }
-
-  const debounceSearch = debounce((event) => handleFilterChange('title_like', event.target.value), 500);
-  // Trigger search
-  searchInput.addEventListener('input', debounceSearch);
-}
-
+// MAIN
 ;(async () => {
   try {
+    const url = new URL(window.location);
+
+    // Update search params if needed
+    if (!url.searchParams.get('_page')) url.searchParams.set('_page', 1);
+    if (!url.searchParams.get('_limit')) url.searchParams.set('_limit', 6);
+
+    history.pushState({}, '', url);
+    const queryParams = url.searchParams;
+
     // Attach click event for links
-    initPagination();
+    initPagination({
+      elementId: 'pagination',
+      defaultParams: queryParams,
+      onChange: (page) => handleFilterChange('_page', page),
+    });
 
-    initSearch();
+    // Search with title_like
+    initSearch({
+      elementId: 'searchInput',
+      defaultParams: queryParams,
+      onChange: (value) => handleFilterChange('title_like', value),
+    });
 
-    // Set default pagination (_page, _limit) on URL
-    initURL();
-
-    // Render post list based URL params
-    const queryParams = new URLSearchParams(window.location.search);
     const { data, pagination } = await postApi.getAll(queryParams);
-    renderPostList(data);
-    renderPagination(pagination);
+    
+    // Render post list based URL params
+    renderPostList('postList', data);
+    renderPagination('pagination', pagination);
   } catch (error) {
     console.log('Get all failed - ' + error);
   }
