+++ js/utils/pagination.js
@@ -1,0 +1,52 @@
+export function renderPagination(elementId, pagination) {
+  const ulPagination = document.getElementById(elementId);
+  if (!pagination || !ulPagination) return;
+
+  // Calc totalPages
+  const { _page, _limit, _totalRows } = pagination;
+  const totalPages = Math.ceil(_totalRows / _limit);
+
+  // Save page and totalPages to ulPagination
+  ulPagination.dataset.page = _page;
+  ulPagination.dataset.totalPages = totalPages;
+
+  // Check if enable/disable prev links
+  if (_page <= 1) ulPagination.firstElementChild?.classList.add('disabled');
+  else ulPagination.firstElementChild?.classList.remove('disabled');
+
+  // Check if enable/disable next links
+  if (_page >= totalPages) ulPagination.lastElementChild?.classList.add('disabled');
+  else ulPagination.lastElementChild?.classList.remove('disabled');
+}
+
+export function initPagination({ elementId, defaultParams, onChange }) {
+  // Bind click event for prev/next link (ul > li > a)
+  const ulPagination = document.getElementById(elementId);
+  if (!ulPagination) return;
+
+  // Set current active page
+
+  // Add click event for prev link
+  const prevLink = ulPagination.firstElementChild?.firstElementChild;
+  if (prevLink) {
+    prevLink.addEventListener('click', (event) => {
+      event.preventDefault();
+
+      // Convert (typeof page) string to number
+      const page = Number.parseInt(ulPagination.dataset.page) || 1;
+      if (page > 2) onChange?.(page - 1);
+    });
+  }
+
+  // Add click event for next link
+  const nextLink = ulPagination.lastElementChild?.lastElementChild;
+  if (nextLink) {
+    nextLink.addEventListener('click', (event) => {
+      event.preventDefault();
+
+      const page = Number.parseInt(ulPagination.dataset.page) || 1;
+      const totalPages = ulPagination.dataset.totalPages;
+      if (page < totalPages) onChange?.(page + 1);
+    });
+  }
+}
