+++ js/utils/search.js
@@ -1,0 +1,17 @@
+import debounce from 'lodash.debounce';
+
+// Pure func - dump func
+export function initSearch({ elementId, defaultParams, onChange }) {
+  const searchInput = document.getElementById(elementId);
+  if (!searchInput) return;
+
+  // Set default values from query params
+  if (defaultParams && defaultParams.get('title_like')) {
+    searchInput.value = defaultParams.get('title_like');
+  }
+
+  const debounceSearch = debounce((event) => onChange?.(event.target.value), 500);
+
+  // Trigger search
+  searchInput.addEventListener('input', debounceSearch);
+}
